+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/ChatAdministratorRights.kt
@@ -1,5 +1,8 @@
 package dev.inmo.tgbotapi.types.chat.member
 
+import dev.inmo.tgbotapi.types.*
+import kotlinx.serialization.SerialName
+
 sealed interface SpecialChatAdministratorRights {
     val canChangeInfo: Boolean
     val canInviteUsers: Boolean
@@ -19,4 +22,40 @@ sealed interface ChatAdministratorRights : SpecialChatAdministratorRights {
     val canPostStories: Boolean
     val canEditStories: Boolean
     val canDeleteStories: Boolean
+
+    companion object {
+        operator fun invoke(
+            canChangeInfo: Boolean = false,
+            canPostMessages: Boolean = false,
+            canEditMessages: Boolean = false,
+            canRemoveMessages: Boolean = false,
+            canInviteUsers: Boolean = false,
+            canRestrictMembers: Boolean = false,
+            canPinMessages: Boolean = false,
+            canPromoteMembers: Boolean = false,
+            canManageVideoChats: Boolean = false,
+            canManageChat: Boolean = false,
+            isAnonymous: Boolean = false,
+            canManageTopics: Boolean = false,
+            canPostStories: Boolean = false,
+            canEditStories: Boolean = false,
+            canDeleteStories: Boolean = false
+        ) = ChatAdministratorRightsImpl(
+            canChangeInfo = canChangeInfo,
+            canPostMessages = canPostMessages,
+            canEditMessages = canEditMessages,
+            canRemoveMessages = canRemoveMessages,
+            canInviteUsers = canInviteUsers,
+            canRestrictMembers = canRestrictMembers,
+            canPinMessages = canPinMessages,
+            canPromoteMembers = canPromoteMembers,
+            canManageVideoChats = canManageVideoChats,
+            canManageChat = canManageChat,
+            isAnonymous = isAnonymous,
+            canManageTopics = canManageTopics,
+            canPostStories = canPostStories,
+            canEditStories = canEditStories,
+            canDeleteStories = canDeleteStories
+        )
+    }
 }